+++ src/app/components/members-collection.js
@@ -4,6 +4,7 @@
  *
  * @author Michał Borzęcki
  * @copyright (C) 2018-2024 ACK CYFRONET AGH
+ * @copyright (C) 2025 Onedata (onedata.org)
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 
@@ -38,6 +39,7 @@ import joinStrings from 'onedata-gui-common/utils/i18n/join-strings';
 import PromiseObject from 'onedata-gui-common/utils/ember/promise-object';
 import ArrayPaginator from 'onedata-gui-common/utils/array-paginator';
 import globals from 'onedata-gui-common/utils/globals';
+import BatchRecordsLoader from 'onezone-gui/utils/batch-records-loader';
 
 const fallbackActionsGenerator = () => [];
 
@@ -51,6 +53,7 @@ export default Component.extend(I18n, {
   currentUser: service(),
   recordManager: service(),
   store: service(),
+  batchRequestRegistry: service(),
 
   /**
    * @override
@@ -250,6 +253,11 @@ export default Component.extend(I18n, {
   searchQuery: undefined,
 
   /**
+   * @type {Map<string, BatchRecordsLoader>}
+   */
+  batchRecordsLoaders: undefined,
+
+  /**
    * @type {string}
    */
   membersTypeText: computed('listHeader', function membersTypeText() {
@@ -305,21 +313,6 @@ export default Component.extend(I18n, {
   ),
 
   /**
-   * Direct groups
-   * @type {Ember.ComputedProperty<PromiseArray<DS.ManyArray<GraphSingleModel>>>}
-   */
-  directGroupsProxy: computed(
-    'record',
-    'subjectType',
-    function directGroupsProxy() {
-      if (this.subjectType === 'group') {
-        return this.directMembersProxy;
-      }
-      return this.getMembers('groupList');
-    }
-  ),
-
-  /**
    * Effective members
    * @type {Ember.ComputedProperty<PromiseArray<DS.ManyArray<GraphSingleModel>>>}
    */
@@ -340,12 +333,10 @@ export default Component.extend(I18n, {
   allMembersLoadingProxy: computed(
     'directMembersProxy',
     'effectiveMembersProxy',
-    'directGroupsProxy',
     function allMembersLoadingProxy() {
       return promiseObject(allFulfilled([
         this.directMembersProxy,
         this.effectiveMembersProxy,
-        this.directGroupsProxy,
       ]));
     }
   ),
@@ -505,11 +496,43 @@ export default Component.extend(I18n, {
     }
   ),
 
-  effectiveMembers: reads('effectiveMemebersProxy.content'),
-
   directMembers: reads('directMembersProxy.content'),
 
   members: reads('membersProxy.content'),
+
+  /**
+   * @type {ComputedProperty<PromiseObject<BatchRecordsLoader>>}
+   */
+  batchRecordsLoaderProxy: computed(
+    'record',
+    'subjectType',
+    function batchRecordsLoaderProxy() {
+      const listName = `eff${_.upperFirst(this.subjectType)}List`;
+      const batchRecordsLoader = this.getBatchRecordsLoader(listName);
+      return promiseObject(batchRecordsLoader);
+    }
+  ),
+
+  /**
+   * @type {ComputedProperty<BatchRecordsLoader>}
+   */
+  batchRecordsLoader: reads('batchRecordsLoaderProxy.content'),
+
+  areRecordsLoadedProxy: computed(
+    'batchRecordsLoader',
+    function areRecordsLoadedProxy() {
+      return promiseObject(this.batchRecordsLoader.getPromise());
+    }),
+
+  /**
+   * @type {ComputedProperty<ProgressTracker>}
+   */
+  progressTracker: computed(
+    'batchRecordsLoader',
+    function progressTracker() {
+      return this.batchRecordsLoader.progressTracker;
+    }
+  ),
 
   membersObserver: observer(
     'members.@each.{entityId,name,username}',
@@ -685,6 +708,7 @@ export default Component.extend(I18n, {
 
   init() {
     this._super(...arguments);
+    this.set('batchRecordsLoaders', new Map());
     this.membersObserver();
     this.groupsObserver();
     this.set('privilegesRecordProxyCache', []);
@@ -725,11 +749,18 @@ export default Component.extend(I18n, {
    */
   getMembers(listName) {
     const record = this.record;
+    const effListName = listName.startsWith('eff') ?
+      listName : `eff${_.upperFirst(listName)}`;
     let promise;
     if (get(record, 'hasViewPrivilege') !== false) {
-      promise = get(record, listName).then(sgl =>
-        sgl ? get(sgl, 'list') : A()
-      );
+      promise = (async () => {
+        const batchRecordsLoader = await this.getBatchRecordsLoader(effListName);
+        await batchRecordsLoader.getPromise();
+
+        return get(record, listName).then(sgl =>
+          sgl ? get(sgl, 'list') : A()
+        );
+      })();
     } else {
       promise = reject({ id: 'forbidden' });
     }
@@ -767,6 +798,56 @@ export default Component.extend(I18n, {
       griAspectPrefix + griAspect,
       subjectId
     );
+  },
+
+  /**
+   * @param {string} listName
+   * @returns {BatchRecordsLoader}
+   */
+  async createBatchRecordsLoader(listName) {
+    const {
+      batchRequestRegistry,
+      record,
+    } = this;
+    let listRecord;
+    if (listName === 'effGroupList' && record.entityType === 'group') {
+      listRecord = await record.getRelation('effChildList');
+    } else {
+      listRecord = await record.getRelation(listName);
+    }
+    const itemsGris = listRecord.hasMany('list').ids();
+
+    const listResolver = async () => {
+      try {
+        // Awaiting for list might fail when some single records cannot be fetched,
+        // but we can still try to read list afterwards.
+        await listRecord.list;
+      } catch {
+        console.warn(
+          'MembersCollection.createBatchRecordsLoader: list cannot be fully resolved, some records may be missing'
+        );
+      }
+      return listRecord.list.toArray();
+    };
+    return new BatchRecordsLoader({
+      batchRequestRegistry,
+      itemsGris,
+      listResolver,
+    });
+  },
+
+  /**
+   * @param {string} listName
+   * @returns {PromiseObject<BatchRecordsLoader>}
+   */
+  getBatchRecordsLoader(listName) {
+    if (!this.batchRecordsLoaders.has(listName)) {
+      this.batchRecordsLoaders.set(
+        listName,
+        promiseObject(this.createBatchRecordsLoader(listName))
+      );
+    }
+    return this.batchRecordsLoaders.get(listName);
   },
 
   actions: {