+++ src/app/locales/en/components/members-collection.js
@@ -28,4 +28,5 @@ export default {
   membershipsTip: 'Visualization of effective membership paths, showing via what groups this member inherits membership in this {{entityType}}.',
   perPage: '{{type}} per page',
   filteredResults: 'matching names',
+  loadingFilters: 'Loading {{type}}... {{progress}}',
 };